requirements.txt
pyzbar==0.1.9
pillow==11.3.0